requirements.txt
asgiref==3.8.1
cfgv==3.4.0
distlib==0.3.9
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
identify==2.6.3
nodeenv==1.9.1
platformdirs==4.3.6
pre_commit==4.0.1
psycopg2==2.9.10
PyJWT==2.10.1
python-decouple==3.8
PyYAML==6.0.2
sqlparse==0.5.2
typing_extensions==4.12.2
virtualenv==20.28.0
